Move test utils to non-test file
