+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/tests/cacher_test.go
@@ -19,65 +19,16 @@ package tests
 
 import (
 	"context"
-	"fmt"
 	"testing"
-	"time"
 
-	apiequality "k8s.io/apimachinery/pkg/api/equality"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/fields"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/apiserver/pkg/apis/example"
-	examplev1 "k8s.io/apiserver/pkg/apis/example/v1"
 	"k8s.io/apiserver/pkg/features"
-	"k8s.io/apiserver/pkg/storage"
-	"k8s.io/apiserver/pkg/storage/cacher"
-	etcd3testing "k8s.io/apiserver/pkg/storage/etcd3/testing"
 	storagetesting "k8s.io/apiserver/pkg/storage/testing"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
 	featuregatetesting "k8s.io/component-base/featuregate/testing"
-	"k8s.io/utils/clock"
 )
 
 func init() {
-	metav1.AddToGroupVersion(scheme, metav1.SchemeGroupVersion)
-	utilruntime.Must(example.AddToScheme(scheme))
-	utilruntime.Must(examplev1.AddToScheme(scheme))
-}
-
-// GetPodAttrs returns labels and fields of a given object for filtering purposes.
-func GetPodAttrs(obj runtime.Object) (labels.Set, fields.Set, error) {
-	pod, ok := obj.(*example.Pod)
-	if !ok {
-		return nil, nil, fmt.Errorf("not a pod")
-	}
-	return labels.Set(pod.ObjectMeta.Labels), PodToSelectableFields(pod), nil
-}
-
-// PodToSelectableFields returns a field set that represents the object
-// TODO: fields are not labels, and the validation rules for them do not apply.
-func PodToSelectableFields(pod *example.Pod) fields.Set {
-	// The purpose of allocation with a given number of elements is to reduce
-	// amount of allocations needed to create the fields.Set. If you add any
-	// field here or the number of object-meta related fields changes, this should
-	// be adjusted.
-	podSpecificFieldsSet := make(fields.Set, 5)
-	podSpecificFieldsSet["spec.nodeName"] = pod.Spec.NodeName
-	podSpecificFieldsSet["spec.restartPolicy"] = string(pod.Spec.RestartPolicy)
-	podSpecificFieldsSet["status.phase"] = string(pod.Status.Phase)
-	return AddObjectMetaFieldsSet(podSpecificFieldsSet, &pod.ObjectMeta, true)
-}
-
-func AddObjectMetaFieldsSet(source fields.Set, objectMeta *metav1.ObjectMeta, hasNamespaceField bool) fields.Set {
-	source["metadata.name"] = objectMeta.Name
-	if hasNamespaceField {
-		source["metadata.namespace"] = objectMeta.Namespace
-	}
-	return source
+	InitTestSchema()
 }
 
 func checkStorageInvariants(ctx context.Context, t *testing.T, key string) {
@@ -352,136 +303,4 @@ func TestWatchSemanticInitialEventsExtended(t *testing.T) {
 	store, terminate := testSetupWithEtcdAndCreateWrapper(t)
 	t.Cleanup(terminate)
 	storagetesting.RunWatchSemanticInitialEventsExtended(context.TODO(), t, store)
-}
-
-// ===================================================
-// Test-setup related function are following.
-// ===================================================
-
-type tearDownFunc func()
-
-type setupOptions struct {
-	resourcePrefix string
-	keyFunc        func(runtime.Object) (string, error)
-	indexerFuncs   map[string]storage.IndexerFunc
-	clock          clock.WithTicker
-}
-
-type setupOption func(*setupOptions)
-
-func withDefaults(options *setupOptions) {
-	prefix := "/pods"
-
-	options.resourcePrefix = prefix
-	options.keyFunc = func(obj runtime.Object) (string, error) { return storage.NamespaceKeyFunc(prefix, obj) }
-	options.clock = clock.RealClock{}
-}
-
-func withClusterScopedKeyFunc(options *setupOptions) {
-	options.keyFunc = func(obj runtime.Object) (string, error) {
-		return storage.NoNamespaceKeyFunc(options.resourcePrefix, obj)
-	}
-}
-
-func withSpecNodeNameIndexerFuncs(options *setupOptions) {
-	options.indexerFuncs = map[string]storage.IndexerFunc{
-		"spec.nodeName": func(obj runtime.Object) string {
-			pod, ok := obj.(*example.Pod)
-			if !ok {
-				return ""
-			}
-			return pod.Spec.NodeName
-		},
-	}
-}
-
-func testSetup(t *testing.T, opts ...setupOption) (context.Context, *cacher.TestCacher, tearDownFunc) {
-	ctx, c, _, tearDown := testSetupWithEtcdServer(t, opts...)
-	return ctx, c, tearDown
-}
-
-func testSetupWithEtcdServer(t *testing.T, opts ...setupOption) (context.Context, *cacher.TestCacher, *etcd3testing.EtcdTestServer, tearDownFunc) {
-	server, etcdStorage := newEtcdTestStorage(t, etcd3testing.PathPrefix())
-	// Inject one list error to make sure we test the relist case.
-	wrappedStorage := &storagetesting.StorageInjectingListErrors{
-		Interface: etcdStorage,
-		Errors:    1,
-	}
-
-	c, err := newTestCacher(wrappedStorage, opts...)
-	if err != nil {
-		t.Fatal(err)
-	}
-	ctx := context.Background()
-	terminate := func() {
-		c.Stop()
-		server.Terminate(t)
-	}
-
-	// Since some tests depend on the fact that GetList shouldn't fail,
-	// we wait until the error from the underlying storage is consumed.
-	if err := wait.PollInfinite(100*time.Millisecond, wrappedStorage.ErrorsConsumed); err != nil {
-		t.Fatalf("Failed to inject list errors: %v", err)
-	}
-
-	return ctx, c, server, terminate
-}
-
-func newTestCacher(s storage.Interface, opts ...setupOption) (*cacher.TestCacher, error) {
-	setupOpts := setupOptions{}
-	opts = append([]setupOption{withDefaults}, opts...)
-	for _, opt := range opts {
-		opt(&setupOpts)
-	}
-	config := cacher.Config{
-		Storage:        s,
-		Versioner:      storage.APIObjectVersioner{},
-		GroupResource:  schema.GroupResource{Resource: "pods"},
-		ResourcePrefix: setupOpts.resourcePrefix,
-		KeyFunc:        setupOpts.keyFunc,
-		GetAttrsFunc:   GetPodAttrs,
-		NewFunc:        newPod,
-		NewListFunc:    newPodList,
-		IndexerFuncs:   setupOpts.indexerFuncs,
-		Codec:          codecs.LegacyCodec(examplev1.SchemeGroupVersion),
-		Clock:          setupOpts.clock,
-	}
-	c, err := cacher.NewCacherFromConfig(config)
-	if err != nil {
-		return nil, err
-	}
-	return &cacher.TestCacher{Cacher: c}, nil
-}
-
-func testSetupWithEtcdAndCreateWrapper(t *testing.T, opts ...setupOption) (storage.Interface, tearDownFunc) {
-	_, c, _, tearDown := testSetupWithEtcdServer(t, opts...)
-
-	if err := c.WaitReady(context.TODO()); err != nil {
-		t.Fatalf("unexpected error waiting for the cache to be ready")
-	}
-	return &createWrapper{TestCacher: c}, tearDown
-}
-
-type createWrapper struct {
-	*cacher.TestCacher
-}
-
-func (c *createWrapper) Create(ctx context.Context, key string, obj, out runtime.Object, ttl uint64) error {
-	if err := c.Cacher.Create(ctx, key, obj, out, ttl); err != nil {
-		return err
-	}
-	return wait.PollUntilContextTimeout(ctx, 100*time.Millisecond, wait.ForeverTestTimeout, true, func(ctx context.Context) (bool, error) {
-		currentObj := c.NewFunc()
-		err := c.Cacher.Get(ctx, key, storage.GetOptions{ResourceVersion: "0"}, currentObj)
-		if err != nil {
-			if storage.IsNotFound(err) {
-				return false, nil
-			}
-			return false, err
-		}
-		if !apiequality.Semantic.DeepEqual(currentObj, out) {
-			return false, nil
-		}
-		return true, nil
-	})
 }
